Include device user count in signals (#2697)
* Include device user count in signals

* Use promise.all
