+++ identity-service/src/routes/fp.js
@@ -1,32 +1,9 @@
 const {
   handleResponse, successResponse, errorResponseBadRequest, errorResponse
 } = require('../apiHelpers')
-const { QueryTypes } = require('sequelize')
-
 const models = require('../models')
-
 const { logger } = require('../logging')
-
-const getDeviceIDCountForUserId = async (userId) => {
-  // Get the # of users sharing any visitorID
-  // associated with `userId` on any platform
-  const res = await models.sequelize.query(
-    `select "Fingerprints"."userId"
-     from "Fingerprints"
-     where "visitorId" in (
-      select distinct "visitorId"
-      from "Fingerprints"
-      where "userId" = :userId
-    ) group by "Fingerprints"."userId"`,
-    {
-      replacements: {
-        userId
-      },
-      type: QueryTypes.SELECT
-    }
-  )
-  return res.length
-}
+const { getDeviceIDCountForUserId } = require('../utils/fpHelpers')
 
 module.exports = function (app) {
   app.post('/fp/webhook', handleResponse(async (req) => {
